[CMake] Attempt to fix non-fatal misleading error: finding testingsupport failed
This is an ordering issue: If existing, testingsupport must be linked before the external compiler-rt is configured.
Once compiler-rt adopts the find_package() approach to pull in its dependencies, this will all get much easier/obsolete.

The current situation caused investigations to go in wrong directions repeatedly. If builds failed (for other reasons), Jenkins was reporting this as the cause:

  llvm-config: error: component libraries and shared library

  llvm-config: error: missing: /Users/buildnode/jenkins/workspace/oss-lldb-osx/Ninja-ReleaseAssert/llvm-macosx-x86_64/lib/libLLVMTestingSupport.a
  CMake Warning at cmake/Modules/CompilerRTUtils.cmake:263 (message):
    llvm-config finding testingsupport failed with status 1
  Call Stack (most recent call first):
    CMakeLists.txt:69 (load_llvm_config)